+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.1.3 - 12.2.2016
+**Added:**
+* Added `matb33:collection-hooks`
+* Added `alanning:roles`
+* All users after registration are added to the `user` group.
+
 v0.1.2 - 7.2.2016
 **Changes:**
 * Added username to schema and improved createdAt